Шёл 1000-7 коммит

+++ MultiLisa/MultiLisa.cpp
@@ -10,31 +10,31 @@ int main()
 {
 	setlocale(LC_ALL, "russian");
 
-	int arraySize;	// Размер массива
-	int maxNumA;	// Число A 
-	int k;			// Число k
+	int arraySize;		// Размер массива
+	float maxNumA;		// Число A 
+	float k;			// Число k
 
-	HANDLE hWork;			// Дескриптор потока work
-	HANDLE hMultiElement;	// Дескриптор потока MultiElement
-	HANDLE hSemaphore;		// Дескриптор семафора для потока work
-	HANDLE hBinSemaphore;	// Дескриптор бинарного семафора для потока MultiElement
+	HANDLE hWork;				// Дескриптор потока work
+	HANDLE hMultiElement;		// Дескриптор потока MultiElement
+	HANDLE hSemaphore;			// Дескриптор семафора для потока work
+	HANDLE hBinSemaphore;		// Дескриптор бинарного семафора для потока MultiElement
 	HANDLE hBinSemaphorePrint;	// Дескриптор бинарного семафора для потока MultiElement
 
-	cout << "Введите размерность массива: ";	//
-	cin >> arraySize;							// Задаём размер массива
-	arraySize += 6;								// и
-	int* array = new int[arraySize];			// создаём массив
-	array[0] = arraySize;						// В первом элементе массива указываем размер массива
+	cout << "Введите размерность массива: ";		//
+	cin >> arraySize;								// Задаём размер массива
+	arraySize += 6;									// и
+	float* array = new float[arraySize];			// создаём массив
+	array[0] = arraySize;							// В первом элементе массива указываем размер массива
 
-	cout << "Введите " << arraySize- 6 << " чесел:" << endl;	//
-	for (int i = 6; i < arraySize; i++)						// Вводим числа
-		cin >> array[i];									//
+	cout << "Введите " << arraySize - 6 << " чесел:" << endl;	//
+	for (int i = 6; i < arraySize; i++)							// Вводим числа
+		cin >> array[i];										//
 
-	cout << "Размерность массива: " << arraySize-5 << endl;	//
-	cout << "Массив: ";										//
-	for (int i = 6; i < arraySize; i++)						// Выводим массив
-		cout << array[i] << " ";							//
-	cout << endl;											//
+	cout << "Размерность массива: " << arraySize - 5 << endl;	//
+	cout << "Массив: ";											//
+	for (int i = 6; i < arraySize; i++)							// Выводим массив
+		cout << array[i] << " ";								//
+	cout << endl;												//
 
 	cout << "Введите число k <= " << arraySize - 6 << ": ";		//
 	cin >> array[1];											// Вводим число k
@@ -57,13 +57,13 @@ int main()
 	if (hWork == NULL)											// Создаём поток work
 		return GetLastError();									//
 
-	hBinSemaphore = CreateSemaphore(NULL, 0, 1, NULL);	//
-	if (hBinSemaphore == NULL)							// Создаём бинарный семафор для потока MultiElement
-		return GetLastError();							//
-	array[4] = (int)hBinSemaphore;						// Сохраняем его в массив
+	hBinSemaphore = CreateSemaphore(NULL, 0, 1, NULL);		//
+	if (hBinSemaphore == NULL)								// Создаём бинарный семафор для потока MultiElement
+		return GetLastError();								//
+	array[4] = (int)hBinSemaphore;							// Сохраняем его в массив
 	hBinSemaphorePrint = CreateSemaphore(NULL, 0, 1, NULL);	//
 	if (hBinSemaphorePrint == NULL)							// Создаём бинарный семафор для потока MultiElement
-		return GetLastError();							//
+		return GetLastError();								//
 	array[5] = (int)hBinSemaphorePrint;						// Сохраняем его в массив
 
 	hMultiElement = CreateThread(NULL, 0, MultiElement, (void*)array, 0, NULL);	//
@@ -71,10 +71,10 @@ int main()
 		return GetLastError();													//
 
 	for (int i = 6; i < arraySize; i++) {					//
-		WaitForSingleObject((HANDLE)array[3], INFINITE);	// Ждём сигнал нового элемента в массиве
+		WaitForSingleObject((HANDLE)(int)array[3], INFINITE);	// Ждём сигнал нового элемента в массиве
 		cout << array[i] << " \a" << flush;					// и выводм элемент
 		if (i + 1 == array[1])								// Проверяем количество прошедших элементов
-			ReleaseSemaphore((HANDLE)array[4], 1, NULL);	// Запускаем поток MultiElement
+			ReleaseSemaphore((HANDLE)(int)array[4], 1, NULL);	// Запускаем поток MultiElement
 	}														//
 	cout << endl;											//
 
@@ -83,7 +83,7 @@ int main()
 		cout << array[i] << " ";		//
 	cout << endl;
 
-	ReleaseSemaphore((HANDLE)array[5], 1, NULL);	// Запускаем поток MultiElement
+	ReleaseSemaphore((HANDLE)(int)array[5], 1, NULL);	// Запускаем поток MultiElement
 
 	WaitForSingleObject(hWork, INFINITE);			// Ждём завершения потоков
 	WaitForSingleObject(hMultiElement, INFINITE);	//
@@ -98,58 +98,58 @@ DWORD WINAPI work(LPVOID array) {
 	cout << "Время сна в милисекундах: ";	// Вводим время сна между итерациями
 	cin >> sleepTime;						//
 
-	int* arrayL = (int*)array;			// Приводим LPVOID к int*
-	int reverseArrayNum = arrayL[0];	// 
-	int offset = 6;						//
-	int* tmp = new int[arrayL[0]- 6];	// Создаём временный массив
+	float* arrayL = (float*)array;			// Приводим LPVOID к int*
+	int reverseArrayNum = arrayL[0];		// 
+	int offset = 6;							//
+	float* tmp = new float[arrayL[0] - 6];	// Создаём временный массив
 
 	for (int i = 0; i < arrayL[0] - 6; i++)	// Копируем данные во временный массив 
 		tmp[i] = arrayL[i + 6];				//
 
 	for (int i = 0; i < arrayL[0]- 6; i++) {					//
-		int num = tmp[i];									//
-		if (num < arrayL[2]) {								// Ищем элементы меньше A
-			arrayL[offset] = num;							// Записываем в лево
-			offset++;										//
-			ReleaseSemaphore((HANDLE)arrayL[3], 1, NULL);	// Сообщаем о завершении итерации
-			Sleep(sleepTime);								// Спим указанное время
-		}													//
-	}														//
+		float num = tmp[i];										//
+		if (num < arrayL[2]) {									// Ищем элементы меньше A
+			arrayL[offset] = num;								// Записываем в лево
+			offset++;											//
+			ReleaseSemaphore((HANDLE)(int)arrayL[3], 1, NULL);		// Сообщаем о завершении итерации
+			Sleep(sleepTime);									// Спим указанное время
+		}														//
+	}															//
 
-	for (int i = 0; i < arrayL[0] - 6; i++) {				//
-		int num = tmp[i];									//
-		if (num == arrayL[2]) {								// Проверяем элементы равные A
-			arrayL[offset] = num;							// Записываем
-			offset++;										//
-			ReleaseSemaphore((HANDLE)arrayL[3], 1, NULL);	// Сообщаем о завершении итерации
-			Sleep(sleepTime);								// Спим указанное время
-		}													//
-	}														//
+	for (int i = 0; i < arrayL[0] - 6; i++) {					//
+		float num = tmp[i];										//
+		if (num == arrayL[2]) {									// Проверяем элементы равные A
+			arrayL[offset] = num;								// Записываем
+			offset++;											//
+			ReleaseSemaphore((HANDLE)(int)arrayL[3], 1, NULL);	// Сообщаем о завершении итерации
+			Sleep(sleepTime);									// Спим указанное время
+		}														//
+	}															//
 
-	for (int i = 0; i < arrayL[0] - 6; i++) {				//
-		int num = tmp[i];									//
-		if (num > arrayL[2]) {								// Ищем элементы больше A
-			arrayL[offset] = num;							// Записываем в право
-			offset++;										//
-			ReleaseSemaphore((HANDLE)arrayL[3], 1, NULL);	// Сообщаем о завершении итерации
-			Sleep(sleepTime);								// Спим указанное время
-		}													//
-	}														//
+	for (int i = 0; i < arrayL[0] - 6; i++) {					//
+		float num = tmp[i];										//
+		if (num > arrayL[2]) {									// Ищем элементы больше A
+			arrayL[offset] = num;								// Записываем в право
+			offset++;											//
+			ReleaseSemaphore((HANDLE)(int)arrayL[3], 1, NULL);	// Сообщаем о завершении итерации
+			Sleep(sleepTime);									// Спим указанное время
+		}														//
+	}															//
 
 	return 0;
 }
 
 DWORD WINAPI MultiElement(LPVOID array) {
-	int* arrayL = (int*)array;	// Приводим LPVOID к int*
-	int k = arrayL[1];		//
+	float* arrayL = (float*)array;	// Приводим LPVOID к int*
+	int k = arrayL[1];			//
 
-	WaitForSingleObject((HANDLE)arrayL[4], INFINITE);	// Ждём сигнал
+	WaitForSingleObject((HANDLE)(int)arrayL[4], INFINITE);	// Ждём сигнал
 
-	int res = 1;					//
+	float res = 1;					//
 	for (int i = 6; i < k; i++)	// Произведение
 		res *= arrayL[i];			//
 
-	WaitForSingleObject((HANDLE)arrayL[5], INFINITE);	// Ждём сигнал
+	WaitForSingleObject((HANDLE)(int)arrayL[5], INFINITE);	// Ждём сигнал
 	cout << "Итоговое произведение: " << res << endl;	// Выводм итоговое произведение
 	return 0;
 }
